Deployment switch (#280)
* Fix type

* Clear db on install

* Logging to external dir

* logging

* Log to Slack

* Change .env

* remove type hint

* test

* Configure SES

* SES dependency

* Change cron

* depl suffix

+++ scripts/install.sh
@@ -19,16 +19,23 @@ if [ ! -v TRAVIS ]; then
 fi
 
 crontab -u ${INSTALLATION_USER} -r
-supervisorctl stop adselect
-supervisorctl stop adpay
 
-./artisan migrate:fresh
-./artisan db:seed
-mongo --eval 'db.dropDatabase()' adselect
-mongo --eval 'db.dropDatabase()' adpay
+if [[ ${DO_RESET} -eq 1 ]]
+then
+    supervisorctl stop adselect${DEPLOYMENT_SUFFIX}
+    supervisorctl stop adpay${DEPLOYMENT_SUFFIX}
 
-supervisorctl start adselect
-supervisorctl start adpay
+    ./artisan migrate:fresh
+    ./artisan db:seed
+
+    mongo --eval 'db.dropDatabase()' adselect${DEPLOYMENT_SUFFIX}
+    mongo --eval 'db.dropDatabase()' adpay${DEPLOYMENT_SUFFIX}
+
+    supervisorctl start adselect${DEPLOYMENT_SUFFIX}
+    supervisorctl start adpay${DEPLOYMENT_SUFFIX}
+else
+    ./artisan migrate
+fi
 
 #./artisan ops:targeting-options:update
 ./artisan ads:fetch-hosts
